+++ src/super_image/models/__init__.py
@@ -30,3 +30,7 @@ from .han.configuration_han import HanConfig
 from .han.modeling_han import HanModel
 from .awsrn.configuration_awsrn import AwsrnConfig
 from .awsrn.modeling_awsrn import AwsrnModel
+from .rnan.configuration_rnan import RnanConfig
+from .rnan.modeling_rnan import RnanModel
+from .ddbpn.configuration_ddbpn import DdbpnConfig
+from .ddbpn.modeling_ddbpn import DdbpnModel