+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- New model: `DdbpnModel`
+- New model: `RnanModel`
 
 ## [0.1.6] - 2021-09-06
 ### Updated